core: Eliminate TOCTTOU races from __setup_folder helper
This isn’t strictly necessary unless two parallel hotdoc instances end
up creating the same directory structure, which would be a bug in
itself.

However, eliminating TOCTTOU races is generally an improvement anyway,
as it reduces the number of system calls and means the code is race-free
if it //is// ever used in parallel.